+++ client/src/sentry.js
@@ -2,6 +2,7 @@
 //
 // SPDX-License-Identifier: GPL-3.0-or-later
 
+// biome-ignore lint/correctness/noUndeclaredDependencies: is defined
 import { init, showReportDialog } from "@sentry/browser";
 
 // Expose only what is needed