+++ weblate/static/editor/base.js
@@ -7,6 +7,8 @@
 var WLT = WLT || {};
 
 WLT.Config = (() => ({
+  // biome-ignore lint/performance/useTopLevelRegex: TODO: factor out
+  // biome-ignore lint/style/useNamingConvention: TODO: fix naming
   IS_MAC: /Mac|iPod|iPhone|iPad/.test(navigator.platform),
 }))();
 
@@ -26,6 +28,20 @@ WLT.Utils = (() => ({
     /* Review workflow */
     $el.find('input[name="review"][value="20"]').prop("checked", true);
   },
+
+  /**
+   * Indicate that the translation has changed
+   * by appending a warning before the editor.
+   * @returns {void}
+   */
+  indicateChanges: () => {
+    const $warning = $("<span class='text-warning'/>");
+    $warning.text(gettext("Unsaved changes!"));
+    if ($(".translation-editor").next(".text-warning").length === 0) {
+      $warning.insertAfter($(".translation-editor"));
+      $(".translation-editor").addClass("has-changes");
+    }
+  },
 }))();
 
 WLT.Editor = (() => {
@@ -41,6 +57,7 @@ WLT.Editor = (() => {
 
     this.$editor.on("input", translationAreaSelector, (e) => {
       WLT.Utils.markTranslated($(e.target).closest("form"));
+      WLT.Utils.indicateChanges();
       hasChanges = true;
     });
 
@@ -105,6 +122,7 @@ WLT.Editor = (() => {
         });
       }
       WLT.Utils.markFuzzy($this.closest("form"));
+      WLT.Utils.indicateChanges();
       hasChanges = true;
       return false;
     });
@@ -117,6 +135,7 @@ WLT.Editor = (() => {
 
       container.find(".translation-editor").attr("dir", direction);
       container.find(".highlighted-output").attr("dir", direction);
+      WLT.Utils.indicateChanges();
       hasChanges = true;
     });
 
@@ -130,6 +149,7 @@ WLT.Editor = (() => {
         .find(".translation-editor")
         .insertAtCaret(text);
       e.preventDefault();
+      WLT.Utils.indicateChanges();
       hasChanges = true;
     });
 
@@ -137,6 +157,15 @@ WLT.Editor = (() => {
     this.init();
 
     this.$translationArea[0].focus();
+
+    // Show confirmation dialog if changes have been made
+    // when leaving the page
+    addEventListener("beforeunload", (e) => {
+      if (hasChanges) {
+        e.preventDefault();
+        return true; // Backwards compatibility
+      }
+    });
 
     // Skip confirmation
     this.$editor.on("click", ".skip", (_e) => {
@@ -148,6 +177,7 @@ WLT.Editor = (() => {
     });
   }
 
+  // biome-ignore lint/suspicious/noEmptyBlockStatements: TODO
   EditorBase.prototype.init = () => {};
 
   EditorBase.prototype.initHighlight = function () {
@@ -159,6 +189,7 @@ WLT.Editor = (() => {
       const $this = $(this);
       insertEditor(this.getAttribute("data-value"), $this);
       e.preventDefault();
+      WLT.Utils.indicateChanges();
       hasChanges = true;
     });
 
@@ -239,6 +270,7 @@ WLT.Editor = (() => {
   }
 
   return {
+    // biome-ignore lint/style/useNamingConvention: TODO
     Base: EditorBase,
   };
 })();