+++ weblate/trans/models/unit.py
@@ -501,7 +501,7 @@ class Unit(models.Model, LoggerMixin):
         # Data for machinery integration
         self.machinery: UnitMemoryResultDict = {}
         # PluralMapper integration
-        self.plural_map: list[str] | None = None
+        self.plural_map: list[str] = []
         # Data for glossary integration
         self.glossary_terms: list[Unit] | None = None
         self.glossary_positions: tuple[tuple[int, int], ...] = ()
@@ -1772,10 +1772,19 @@ class Unit(models.Model, LoggerMixin):
 
     def get_secondary_units(self, user: User) -> list[Unit]:
         """Return list of secondary units."""
-        secondary_langs: set[int] = user.profile.secondary_language_ids - {
-            self.translation.language_id,
-            self.translation.component.source_language_id,
-        }
+        translation = self.translation
+        component = translation.component
+        secondary_langs: set[int] = user.profile.secondary_language_ids
+
+        # Add project/component secondary languages
+        if component.secondary_language_id:
+            secondary_langs.add(component.secondary_language_id)
+        elif component.project.secondary_language_id:
+            secondary_langs.add(component.project.secondary_language_id)
+
+        # Remove current source and trarget language
+        secondary_langs -= {translation.language_id, component.source_language_id}
+
         if not secondary_langs:
             return []
         result = get_distinct_translations(
@@ -1791,8 +1800,9 @@ class Unit(models.Model, LoggerMixin):
                 "translation__plural",
             )
         )
+        # Avoid fetching component again from the database
         for unit in result:
-            unit.translation.component = self.translation.component
+            unit.translation.component = component
         return result
 
     @property
@@ -1959,10 +1969,12 @@ class Unit(models.Model, LoggerMixin):
 
     def update_explanation(
         self, explanation: str, user: User, save: bool = True
-    ) -> None:
+    ) -> bool:
         """Update glossary explanation."""
         verify_in_transaction()
-        old = self.explanation
+        old = self.old_unit["explanation"]
+        if old == explanation:
+            return False
         self.explanation = explanation
         file_format_support = (
             self.translation.component.file_format_cls.supports_explanation
@@ -1989,12 +2001,13 @@ class Unit(models.Model, LoggerMixin):
                 target=explanation,
                 old=old,
             )
+        return True
 
     def update_extra_flags(
         self, extra_flags: str, user: User, save: bool = True
     ) -> None:
         """Update unit extra flags."""
-        old = self.extra_flags
+        old = self.old_unit["extra_flags"]
         self.extra_flags = extra_flags
         units: Iterable[Unit] = []
         if self.is_source: